refactor: rename brand to logo
